// src/app/models/model-docente-busqueda.ts
import { ModelMensaje } from "./model-mensaje";

export class ModelDocenteBusqueda extends ModelMensaje {
    ASODNI?: string;
    ASOAPENOMDNI?: string;
    ASOFECNAC?: string;
    ASODIR?: string;
    ASOTELF1?: string;
    ASOTELF2?: string;
    ASOEMAIL?: string;
    ASOTIPID?: string;
    TIPO_ASOCIADO?: string;
    SEXO?: string;
    REGPENID?: string;
    ASOFRESNOM?: string;
    ASOFRESCESE?: string;
    AUTDESAPO?: string;
    ASOSEXO?: string;
    REGPENDES?: string;
    AFILIADO?: string;
    ESTCIVID?: string;
    ESTCIVDES?: string;
    ESTCIVDESFEM?: string;
    NOMCENEDU?: string;
    DEPARTAMENTO?: string;
    PROVINCIA?: string;
    DISTRITO?: string;
    FOTO?: string;
    FIRMA?: string;
    FALLECIDO?: string;
    TOTAL?: number;
    N?: number;
}